requirements.txt
numpy
biopython
python-Levenshtein
scikit-learn
matplotlib
seaborn
torch==1.4.0